Fix linting

+++ math/include/rawrbox/math/color.hpp
@@ -17,7 +17,6 @@ namespace rawrBox {
 		Color_t() = default;
 		Color_t(NumberType _r, NumberType _g, NumberType _b, NumberType _a) : r(_r), g(_g), b(_b), a(_a) {}
 
-		// NOLINTBEGIN(clang-diagnostic-c++11-narrowing)
 		static Color_t<NumberType> RGBAHex(uint32_t x) {
 			if constexpr (std::is_same<NumberType, float>::value) {
 				return {
@@ -43,13 +42,12 @@ namespace rawrBox {
 				    1.F};
 			} else {
 				return {
-				    ((x >> 16) & 0xFF),
-				    ((x >> 8) & 0xFF),
-				    ((x)&0xFF),
+				    static_cast<int>(((x >> 16) & 0xFF)),
+				    static_cast<int>(((x >> 8) & 0xFF)),
+				    static_cast<int>(((x)&0xFF)),
 				    255};
 			}
 		}
-		// NOLINTEND(clang-diagnostic-c++11-narrowing)
 
 		static uint32_t toHEX(Color_t<NumberType> c) {
 			if constexpr (std::is_same<NumberType, int>::value) {
